+++ GaluhadPlugins/Pets/Strings/French/Pets.lua
@@ -464,11 +464,17 @@ _PETSTRINGS[381] = {[1]="Serpent vert vif";[2]="Offre à durée limitée : suiv
 _PETSTRINGS[382] = {[1]="Adorable ver des sables";[2]="Disponible par échange en participant à des excavations."}
 _PETSTRINGS[383] = {[1]="Petit-gris de la courbe parfaite";[2]="Disponible par échange au cours du Festival du printemps."}
 _PETSTRINGS[384] = {[1]="Crabe";[2]="Disponible par échange en participant à des excavations."}
-_PETSTRINGS[385] = {[1]="Macaque rusé";[2]="Disponible dans l'Instance L'île des tempêtes."}
+_PETSTRINGS[385] = {[1]="Macaque rusé";[2]="Disponible dans l'Instance Île des tempêtes."}
 _PETSTRINGS[386] = {[1]="Sitaka";[2]="Disponibilité limitée pour les Mirages de splendeur et dans les coffres à butin de voyageur et d'aventurier."}
 _PETSTRINGS[387] = {[1]="Créatures fantaisistes de la forêt";[2]="Obtenu dans le Coffre luxuriant du client fantaisiste."}
 _PETSTRINGS[388] = {[1]="Hamster brun";[2]="Disponible par échange au cours du Festival d'anniversaire."}
 _PETSTRINGS[389] = {[1]="Hamster jovial";[2]="Disponible par échange au cours du Festival d'anniversaire."}
+
+_PETSTRINGS[390] = {[1]="Gerboise albinos";[2]="Peut être échangé auprès de l'intendant des récompenses de la cité d'Umbar."}
+_PETSTRINGS[391] = {[1]="Gerboise gris ardoise";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 10."}
+_PETSTRINGS[392] = {[1]="Perroquet d'Umbar Baharbêl";[2]="S'obtient en accomplissant des missions à Umbar Baharbêl."}
+_PETSTRINGS[393] = {[1]="Lionne altière";[2]="Peut être échangé auprès de l'intendant des récompenses de la cité d'Umbar."}
+_PETSTRINGS[394] = {[1]="Lion à crinière";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 10."}
 
 _PETSTRINGS[500] = {[1]="Chauve-souris rayée";[2]=""}
 _PETSTRINGS[501] = {[1]="Chauve-souris blessée";[2]=""}
@@ -488,20 +494,26 @@ _PETSTRINGS[514] = {[1]="Aurochs automnal";[2]="";}
 _PETSTRINGS[515] = {[1]="Aurochs fauve";[2]="";}
 _PETSTRINGS[516] = {[1]="Gerboise";[2]="";}
 _PETSTRINGS[517] = {[1]="Gerboise à rayures";[2]=""}
-_PETSTRINGS[518] = {[1]="Gerboise blanche";[2]=""}
-_PETSTRINGS[519] = {[1]="Gerboise brune";[2]=""}
-_PETSTRINGS[520] = {[1]="Gerboise noire";[2]=""}
-_PETSTRINGS[521] = {[1]="Corgi blond";[2]=""}
-_PETSTRINGS[522] = {[1]="Serpent bordeaux";[2]=""}
-_PETSTRINGS[523] = {[1]="Serpent moucheté de jade";[2]=""}
-_PETSTRINGS[524] = {[1]="Serpent turquoise";[2]="";}
-_PETSTRINGS[525] = {[1]="Serpent tacheté";[2]="";}
-_PETSTRINGS[526] = {[1]="Serpent vert forêt";[2]="";}
-_PETSTRINGS[527] = {[1]="Serpent vert émeraude";[2]=""}
-_PETSTRINGS[528] = {[1]="Serpent ardent du désert";[2]=""}
-_PETSTRINGS[529] = {[1]="Ver des sables vermeil";[2]=""}
-_PETSTRINGS[530] = {[1]="Ver des sables châtain";[2]=""}
-_PETSTRINGS[531] = {[1]="Hamster moucheté";[2]=""}
-_PETSTRINGS[532] = {[1]="Hamster blanc";[2]=""}
-_PETSTRINGS[533] = {[1]="Hamster taupe";[2]=""}
-_PETSTRINGS[534] = {[1]="Hamster mauve";[2]=""}
+_PETSTRINGS[518] = {[1]="Gerboise brune";[2]=""}
+_PETSTRINGS[519] = {[1]="Corgi blond";[2]=""}
+_PETSTRINGS[520] = {[1]="Serpent bordeaux";[2]=""}
+_PETSTRINGS[521] = {[1]="Serpent moucheté de jade";[2]=""}
+_PETSTRINGS[522] = {[1]="Serpent turquoise";[2]=""}
+_PETSTRINGS[523] = {[1]="Serpent tacheté";[2]=""}
+_PETSTRINGS[524] = {[1]="Serpent vert forêt";[2]="";}
+_PETSTRINGS[525] = {[1]="Serpent vert émeraude";[2]="";}
+_PETSTRINGS[526] = {[1]="Serpent ardent du désert";[2]="";}
+_PETSTRINGS[527] = {[1]="Ver des sables vermeil";[2]=""}
+_PETSTRINGS[528] = {[1]="Ver des sables châtain";[2]=""}
+_PETSTRINGS[529] = {[1]="Hamster moucheté";[2]=""}
+_PETSTRINGS[530] = {[1]="Hamster blanc";[2]=""}
+_PETSTRINGS[531] = {[1]="Hamster taupe";[2]=""}
+_PETSTRINGS[532] = {[1]="Hamster mauve";[2]=""}
+_PETSTRINGS[533] = {[1]="Perroquet des sables ardents";[2]=""}
+_PETSTRINGS[534] = {[1]="Bufflon d'eau";[2]=""}
+_PETSTRINGS[535] = {[1]="Bufflonne d'eau";[2]=""}
+_PETSTRINGS[536] = {[1]="Perroquet de lumière";[2]=""}
+_PETSTRINGS[537] = {[1]="Buffle d'eau";[2]=""}
+_PETSTRINGS[538] = {[1]="Perroquet des ténèbres";[2]=""}
+_PETSTRINGS[539] = {[1]="Perroquet mat";[2]=""}
+_PETSTRINGS[540] = {[1]="Perroquet mat fringant";[2]=""}