+++ GaluhadPlugins/Pets/Strings/German/Pets.lua
@@ -466,6 +466,12 @@ _PETSTRINGS[387] = {[1]="Wunderliche Waldtiere";[2]="Aus einer luxuriösen Truhe
 _PETSTRINGS[388] = {[1]="Sienafarbener Hamster";[2]="Durch Tauschhandel während des Jubiläumsfests erhältlich"}
 _PETSTRINGS[389] = {[1]="Fröhlicher Hamster";[2]="Durch Tauschhandel während des Jubiläumsfests erhältlich"}
 
+_PETSTRINGS[390] = {[1]="Albino-Springmaus";[2]="Zum Tausch verfügbar beim Belohnungs-Quartiermeister der Stadt Umbar in Umbar."}
+_PETSTRINGS[391] = {[1]="Schiefergraue Springmaus";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 10"}
+_PETSTRINGS[392] = {[1]="Papagei von Umbar Baharbêl";[2]="Erhältlich durch den Abschluss von Taten in Umbar Baharbêl."}
+_PETSTRINGS[393] = {[1]="Stolze Löwin";[2]="Zum Tausch verfügbar beim Belohnungs-Quartiermeister der Stadt Umbar in Umbar."}
+_PETSTRINGS[394] = {[1]="Bemähnter Löwe";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 10"}
+
 _PETSTRINGS[500] = {[1]="Gestreifte Fledermaus";[2]=""}
 _PETSTRINGS[501] = {[1]="Verletzte Fledermaus";[2]=""}
 _PETSTRINGS[502] = {[1]="Hellbraune Fledermaus";[2]=""}
@@ -484,20 +490,26 @@ _PETSTRINGS[514] = {[1]="Herbstroter Auerochse";[2]="";}
 _PETSTRINGS[515] = {[1]="Lohfarbener Auerochse";[2]="";}
 _PETSTRINGS[516] = {[1]="Springmaus";[2]="";}
 _PETSTRINGS[517] = {[1]="Gestreifte Springmaus";[2]=""}
-_PETSTRINGS[518] = {[1]="Weiße Springmaus";[2]=""}
-_PETSTRINGS[519] = {[1]="Braune Springmaus";[2]=""}
-_PETSTRINGS[520] = {[1]="Schwarze Springmaus";[2]=""}
-_PETSTRINGS[521] = {[1]="Heller Corgi";[2]=""}
-_PETSTRINGS[522] = {[1]="Weinrote Schlange";[2]=""}
-_PETSTRINGS[523] = {[1]="Gefleckte Jadeschlange";[2]=""}
-_PETSTRINGS[524] = {[1]="Türkise Schlange";[2]="";}
-_PETSTRINGS[525] = {[1]="Gefleckte Schlange";[2]="";}
-_PETSTRINGS[526] = {[1]="Blattgrüne Schlange";[2]="";}
-_PETSTRINGS[527] = {[1]="Smaragdgrüne Schlange";[2]=""}
-_PETSTRINGS[528] = {[1]="Feurige Wüstenschlange";[2]=""}
-_PETSTRINGS[529] = {[1]="Rubinroter Sandwurm";[2]=""}
-_PETSTRINGS[530] = {[1]="Kastanienfarbener Sandwurm";[2]=""}
-_PETSTRINGS[531] = {[1]="Gefleckter Hamster";[2]=""}
-_PETSTRINGS[532] = {[1]="Weißer Hamster";[2]=""}
-_PETSTRINGS[533] = {[1]="Graubrauner Hamster";[2]=""}
-_PETSTRINGS[534] = {[1]="Heidehamster";[2]=""}
+_PETSTRINGS[518] = {[1]="Braune Springmaus";[2]=""}
+_PETSTRINGS[519] = {[1]="Heller Corgi";[2]=""}
+_PETSTRINGS[520] = {[1]="Weinrote Schlange";[2]=""}
+_PETSTRINGS[521] = {[1]="Gefleckte Jadeschlange";[2]=""}
+_PETSTRINGS[522] = {[1]="Türkise Schlange";[2]=""}
+_PETSTRINGS[523] = {[1]="Gefleckte Schlange";[2]=""}
+_PETSTRINGS[524] = {[1]="Blattgrüne Schlange";[2]="";}
+_PETSTRINGS[525] = {[1]="Smaragdgrüne Schlange";[2]="";}
+_PETSTRINGS[526] = {[1]="Feurige Wüstenschlange";[2]="";}
+_PETSTRINGS[527] = {[1]="Rubinroter Sandwurm";[2]=""}
+_PETSTRINGS[528] = {[1]="Kastanienfarbener Sandwurm";[2]=""}
+_PETSTRINGS[529] = {[1]="Gefleckter Hamster";[2]=""}
+_PETSTRINGS[530] = {[1]="Weißer Hamster";[2]=""}
+_PETSTRINGS[531] = {[1]="Graubrauner Hamster";[2]=""}
+_PETSTRINGS[532] = {[1]="Heidehamster";[2]=""}
+_PETSTRINGS[533] = {[1]="Papagei der feurigen Sande";[2]=""}
+_PETSTRINGS[534] = {[1]="Wasserbüffelkalb";[2]=""}
+_PETSTRINGS[535] = {[1]="Wasserbüffelkuh";[2]=""}
+_PETSTRINGS[536] = {[1]="Papagei des Lichts";[2]=""}
+_PETSTRINGS[537] = {[1]="Wasserbüffelochse";[2]=""}
+_PETSTRINGS[538] = {[1]="Papagei der Dunkelheit";[2]=""}
+_PETSTRINGS[539] = {[1]="Dunkler Papagei";[2]=""}
+_PETSTRINGS[540] = {[1]="Eleganter dunkler Papagei";[2]=""}