+++ Changelog.md
@@ -1,4 +1,8 @@
 Change log:
+* 2024-05-16 - U40 - 3 new, 2 newly available, 8 future pets
+  * Parrot of Umbar Baharbêl, Proud Lioness, Maned Lion
+  * Albino Jerboa, Slate Jerboa
+  * Parrot of the Fiery Sands, Water Buffalo Calf, Water Buffalo Cow, Parrot of Light, Water Buffalo Bull, Parrot of Darkness, Dusky Parrot, Dashing Dusky Parrot
 * 2024-04-22 - U39.2 new pet 3 new + 4 future
   * Whimsical Woodland Critters, Sienna Hamster, Merrymaking Hamster
   * Spotted Hamster, White Hamster, Taupe Hamster, Heather Hamster